+++ src/ThreeBody3D.jl
@@ -15,14 +15,16 @@ include("Visualization.jl")
 
 export ThreeBodySystem,
        STATE_SIZE, statevector, body_position, velocity,
-       AccuracyProfile, accuracy_profile, SimulationResult, simulate,
+       AccuracyProfile, accuracy_profile,
+       CloseApproachEvent, SimulationResult, simulate,
+       terminated_by_close_approach,
        center_of_mass, center_of_mass_velocity,
        linear_momentum, angular_momentum,
        kinetic_energy, potential_energy, total_energy,
        minimum_separation, relative_energy_error,
        DiagnosticsReport, diagnostics_report,
        periodicity_error, CloseApproachReport, close_approach_report,
-       SolverBenchmark, benchmark_solvers,
+       SolverBenchmark, benchmark_solvers, benchmark_extreme_solvers,
        plot_trajectory, animate, record_animation
 
 end